backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
holidays==0.71
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
prophet==1.1.6
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
realtime==2.4.3
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stanio==0.5.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==14.2
yarl==1.20.0
